+++ src/onegov.election_day/onegov/election_day/forms/upload/__init__.py
@@ -1,7 +1,7 @@
 from onegov.election_day.forms.upload.election import UploadMajorzElectionForm
 from onegov.election_day.forms.upload.election import UploadProporzElectionForm
 from onegov.election_day.forms.upload.party_results import \
-    UploadElectionPartyResultsForm
+    UploadPartyResultsForm
 from onegov.election_day.forms.upload.rest import UploadRestForm
 from onegov.election_day.forms.upload.vote import UploadVoteForm
 from onegov.election_day.forms.upload.wabsti_majorz import \
@@ -12,7 +12,7 @@ from onegov.election_day.forms.upload.wabsti_vote import UploadWabstiVoteForm
 
 
 __all__ = [
-    'UploadElectionPartyResultsForm',
+    'UploadPartyResultsForm',
     'UploadMajorzElectionForm',
     'UploadProporzElectionForm',
     'UploadRestForm',