+++ src/onegov.election_day/onegov/election_day/layouts/election_compound.py
@@ -1,6 +1,8 @@
 from cached_property import cached_property
+from onegov.core.utils import normalize_for_url
 from onegov.election_day import _
 from onegov.election_day.layouts.detail import DetailLayout
+from onegov.election_day.utils import svg_filename
 
 
 class ElectionCompoundLayout(DetailLayout):
@@ -14,6 +16,7 @@ class ElectionCompoundLayout(DetailLayout):
         return (
             'districts',
             'candidates',
+            'parties',
             'data'
         )
 
@@ -24,6 +27,8 @@ class ElectionCompoundLayout(DetailLayout):
             return self.request.app.principal.label('districts')
         if tab == 'candidates':
             return _("Elected candidates")
+        if tab == 'parties':
+            return _("Parties")
         if tab == 'data':
             return _("Downloads")
 
@@ -32,6 +37,8 @@ class ElectionCompoundLayout(DetailLayout):
     def visible(self, tab=None):
         if not self.has_results:
             return False
+        if tab == 'parties':
+            return self.model.party_results.first()
 
         return True
 
@@ -60,3 +67,33 @@ class ElectionCompoundLayout(DetailLayout):
         if not self.model.elections:
             return False
         return self.model.elections[0].type == 'proporz'
+
+    @cached_property
+    def svg_path(self):
+        """ Returns the path to the SVG or None, if it is not available. """
+
+        path = 'svg/{}'.format(
+            svg_filename(self.model, self.tab)
+        )
+        if self.request.app.filestorage.exists(path):
+            return path
+
+        return None
+
+    @cached_property
+    def svg_link(self):
+        """ Returns a link to the SVG download view. """
+
+        return self.request.link(self.model, name='{}-svg'.format(self.tab))
+
+    @cached_property
+    def svg_name(self):
+        """ Returns a nice to read SVG filename. """
+
+        return '{}.svg'.format(
+            normalize_for_url(
+                '{}-{}'.format(
+                    self.model.id, self.title() or ''
+                )
+            )
+        )